Implemented lasers
Each controller has a laser. Left hand has a blue laser, right hand has a red laser. Lasers can be toggled by pressing a controller's primary button.
I was planning on using scenes for the controllers, but it seemed difficult to set the colors with scenes. I could use the same scripts, so I felt like that was a good compromise.
One issue is that the rotation and laser width is a bit strange. One of the lasers is almost 2D with one half of a circle on each end of the laser. Rotating either controller seems to rotate the laser. What makes this especially strange is that it seems to always be whatever laser I instantiate second. I tried to fix it by recreating one controller, then it just started happening with the other one. It doesn't seem to cause any errors, it just looks strange.