Created cube spawner
Created a cube spawner node. Spawns a blue or red cube every second that moves towards the player and is destroyed after traveling a certain distance.
Sword collision isn't working. I'm planning to just use cylinders instead of the line renderer since it's making this harder than it has to be.